Player utilizes state machines. Moved InputHandler logic into Player
