+++ Agents/Instances/Player/States/PlayerRunningState.cs
@@ -9,4 +9,28 @@ public partial class PlayerRunningState : State<Player>
 
     public override string Key { get; set; } = "RUNNING";
 
+    public override void Enter()
+    {
+        base.Enter();
+
+        Parent.SetSprinting(true);
+    }
+
+    public override void Exit()
+    {
+        base.Exit();
+
+        Parent.SetSprinting(false);
+    }
+
+    public override void OnProcess(double delta)
+    {
+        base.OnProcess(delta);
+
+        if (!Input.IsActionPressed("action_sprint"))
+        {
+            ParentStateMachine.TransitionTo(CharacterController.States.Walking.ToString());
+        }
+    }
+
 }
